requirements.txt
PySide6>=6.5.0
pyqtgraph>=0.13.0
praat-parselmouth>=0.4.3
librosa>=0.10.0
numpy>=1.22.0
scipy>=1.8.0
pyinstaller>=6.0.0
soundfile>=0.12.0
